Refactor: Move character data fetch to pages/index.tsx
