+++ types/global.d.ts
@@ -1,7 +1,5 @@
 export type TDashboardProps = {
-  uid: string;
   lodestoneId: number | null;
-  initialResults: any
 };
 
 export type TRoleAccordionProps = {
